chore: cmd pattern refactor, write tests
